requirements.txt
PyYAML==6.0
selenium==4.5.0
webdriver-manager==3.8.3